backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
ultralytics
torch
torchvision
numpy
opencv-python
huggingface-hub
facenet-pytorch
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.2
APScheduler==3.10.4
